requirements.txt
fastapi
numpy
Pillow
pydantic
uvicorn[standard]
gunicorn
tensorflow-cpu
opencv-python-headless
python-multipart
scikit-learn
